Fix dtype conversion bug in scoring script
Same issue as probe model training - scoring script was passing uint16
tensors to the model, but embedding layer requires Long (int64) type.

Error:
  RuntimeError: Expected tensor for argument #1 'indices' to have one
  of the following scalar types: Long, Int; but got CUDAUInt16Type instead

Fix:
- Convert input_ids to torch.long when moving to device
- Convert targets to torch.long when moving to device
- Use .to(device, dtype=torch.long) for efficiency (single operation)

This ensures compute_nll_score() passes correct dtypes to model.

+++ data/score_fineweb_entropy.py
@@ -116,9 +116,9 @@ def compute_nll_score(model: SimpleGPT, document: Tensor, device: torch.device, 
             end = min(start + max_seq_len, doc_len - 1)
             chunk_len = end - start
 
-            # Get input and target
-            input_ids = document[start:end].unsqueeze(0).to(device)  # [1, chunk_len]
-            targets = document[start + 1:end + 1].to(device)  # [chunk_len]
+            # Get input and target (convert uint16 -> int64 for embedding)
+            input_ids = document[start:end].unsqueeze(0).to(device, dtype=torch.long)  # [1, chunk_len]
+            targets = document[start + 1:end + 1].to(device, dtype=torch.long)  # [chunk_len]
 
             # Forward pass
             logits = model(input_ids)  # [1, chunk_len, vocab]
